Refactored Square component to separate styles from business logic and loosely couple to number of boxes.
