+++ src/components/row.component.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "styled-components";
 
-import Square from "./square.component";
+import Square from "./square/square.component";
 
 const RowContainer = styled.div`
     display: grid;
