+++ src/constants.js
@@ -1,0 +1,11 @@
+//** BOARD LAYOUT CONSTANTS: START */
+
+export const BOARD_ORDER = 3;
+
+//** BOARD LAYOUT CONSTANTS: END */
+
+//** CALCULATED CONSTANTS: START */
+
+export const TOTAL_BOXES_PER_ELEMENT = BOARD_ORDER * BOARD_ORDER;
+
+//** CALCULATED CONSTANTS: END */
